requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0

scikit-learn==1.7.2
pandas==2.3.3
python-dotenv==1.0.0